Fix order of iteration over history of items changes
DO NOT FORGET: the natural order of reading the journal is
"from oldest to newest" version of item, so the last/current/actual version of
item is the last occurrence in the array, to find it effectively you should
iterate over array in reverse order. There is only one exception
from that rule it is array of TagMap, for this type of items,
regardless of direction of traversal, all maps "tag: item" are collected.